requirements.txt
# RRF Search Project Dependencies

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Database
## aws RDS postgreSQL 활용

# Machine Learning & NLP
sentence-transformers==5.1.2
numpy==1.24.3
torch==2.3.0
huggingface-hub==0.36.0

# Korean NLP
kiwipiepy==0.16.2
kiwipiepy-model==0.16.0
rank-bm25==0.2.2

# LLM Integration (AWS Bedrock)
anthropic==0.40.0
boto3==1.34.0
botocore==1.34.0

# Utilities
python-dotenv==1.0.0

# Development (Optional)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2